refactor: keep actions to commands, and the wiring where it runs
Two files sat on the wrong side of the line between `actions/` — what the
user does, with its side effects and its single undo entry — and the rest.

`importBPMN` loads imported cells into the graph, normalising the structure
as it goes. That is the document command, and it lived in `utils/` while
the file reading that drives it lived in `actions/`: one feature split
across both folders. It moves next to `importFile`, its only caller, and
stops being exported at all.

`setupFileImport` was the reverse — not a command but effect wiring: it
attached three listeners to the scroller and returned a cleanup, for one
caller that ran it inside `useEffect`. That is what an effect body is, so
it moves into `FileImportOverlay`, which already owned the overlay it
toggles. The component now imports the command instead of its own plumbing.

The handlers are arrow consts rather than declarations: TypeScript carries
the guard's narrowing into a closure but not into a hoisted `function`, so
declarations would have cost six non-null assertions.
